+++ drivers/local_storage/onitu_local_storage/tests/driver.py
@@ -3,12 +3,12 @@ import shutil
 
 from path import path
 
-from tests.utils.testdriver import TestDriver
+from tests.utils import driver
 from tests.utils import files
 from tests.utils.tempdirs import dirs
 
 
-class Driver(TestDriver):
+class Driver(driver.Driver):
     SPEED_BUMP = 1
 
     def __init__(self, *args, **options):
