+++ tests/functional/core/test_copy.py
@@ -1,10 +1,10 @@
 import pytest
 
-from tests.utils.driver import TestingDriver
+from tests.utils.testdriver import TestDriver
 from tests.utils.loop import BooleanLoop
 from tests.utils.files import KB, MB
 
-rep1, rep2 = TestingDriver('rep1'), TestingDriver('rep2')
+rep1, rep2 = TestDriver('rep1'), TestDriver('rep2')
 
 
 @pytest.fixture(autouse=True)
