+++ tests/functional/core/test_corruption.py
@@ -2,13 +2,13 @@ import hashlib
 
 import pytest
 
-from tests.utils.driver import TestingDriver
+from tests.utils.testdriver import TestDriver
 from tests.utils.loop import BooleanLoop
 
 # We use chunks of size 1 to slow down the transfers. This way, we have
 # more chances to stop a transfer before its completion
-rep1 = TestingDriver('rep1')
-rep2 = TestingDriver('rep2', speed_bump=True)
+rep1 = TestDriver('rep1')
+rep2 = TestDriver('rep2', speed_bump=True)
 
 
 @pytest.fixture(autouse=True)
