+++ tests/functional/core/test_move.py
@@ -1,9 +1,9 @@
 import pytest
 
-from tests.utils.driver import TestingDriver
+from tests.utils.testdriver import TestDriver
 from tests.utils.loop import BooleanLoop, CounterLoop
 
-rep1, rep2 = TestingDriver('rep1'), TestingDriver('rep2')
+rep1, rep2 = TestDriver('rep1'), TestDriver('rep2')
 
 
 @pytest.fixture(autouse=True)
