+++ tests/utils/driver.py
@@ -1,46 +1,56 @@
-import os
+import re
 import pkg_resources
 
-from .testdriver import TestDriver
+
+class Driver(object):
+
+    SPEED_BUMP = 0
+    """
+    Can be used by fast drivers (like LocalStorage) to slow down
+    some tests.
+
+    The value is the chunk_size which will be used in the tests
+    requiring slow-downs. If 0, no speed-bumps are used (default).
+    """
+
+    def __init__(self, type, name=None, speed_bump=False, **options):
+        self.type = type
+        self.name = name if name else type
+        self.options = options
+
+        if speed_bump and self.SPEED_BUMP > 0:
+            self.options['chunk_size'] = self.SPEED_BUMP
+
+    @property
+    def slugname(self):
+        return re.sub(r'__+', '_', re.sub(r'[^a-z0-9]', '_',
+                                          self.name.lower()))
+
+    @property
+    def dump(self):
+        return {'driver': self.type, 'options': self.options}
+
+    @property
+    def id(self):
+        return (self.type, self.name)
+
+    def connect(self, session):
+        pass
+
+    def __hash__(self):
+        return hash(self.id)
+
+    def __eq__(self, other):
+        return self.id == other.id
 
 
-class Driver(TestDriver):
-    drivers = {}
-
-    def __new__(cls, name, *args, **kwargs):
-        entry_points = pkg_resources.iter_entry_points('onitu.tests')
-        tests_modules = {e.name: e for e in entry_points}
-
-        if name not in tests_modules:
-            raise ImportError(
-                "Cannot import tests for driver {}".format(name)
-            )
-
-        try:
-            tests = tests_modules[name].load()
-        except ImportError as e:
-            raise ImportError(
-                "Error importing tests for driver {}: {}".format(name, e)
-            )
-
-        try:
-            driver = tests.Driver
-        except ImportError:
-            raise ImportError(
-                "Tests for driver {} don't expose a"
-                "Driver class".format(name)
-            )
-
-        cls.drivers[name] = driver
-        return driver(*args, **kwargs)
-
-
-class TestingDriver(Driver):
-    def __new__(cls, *args, **kwargs):
-        return Driver('test', *args, **kwargs)
-
-
-class TargetDriver(Driver):
-    def __new__(cls, *args, **kwargs):
-        name = os.environ.get('ONITU_TEST_DRIVER', 'test')
-        return Driver(name, *args, **kwargs)
+def load_driver(name):
+    try:
+        entry_point = next(iter(pkg_resources.iter_entry_points('onitu.tests',
+                                                                name)))
+        return entry_point.load().Driver
+    except StopIteration:
+        raise ImportError("Cannot import tests for driver {}".format(name))
+    except Exception as e:
+        raise ImportError("Error importing tests for driver {}:"
+                          "{}: {}".format(name, type(e).__name__, e))
